✨ feat(deletepost): removed post's main image and post's folder on cloudinary while deleting post

+++ src/modules/posts/controllers/deletePost.ts
@@ -4,6 +4,7 @@ import { prisma } from "../../../services/prismaClient.js";
 import { getJsonResponse } from "../../../utils/helperMethods/getJsonResponse.js";
 import { getSuccessMsg } from "../../../utils/helperMethods/generateSuccessMsg.js";
 import { getAuth } from "@clerk/express";
+import cloudinary from "../../../services/cloudinary.js";
 
 export async function deletePost(
   req: Request,
@@ -23,6 +24,22 @@ export async function deletePost(
     const userEmail = user?.email ?? "";
 
     const postSlug: string = req.params.slug;
+
+    if (!postSlug) return next(new Error("Post slug is required."));
+
+    const post = await prisma.post.findUnique({
+      where: {
+        slug: postSlug,
+        userEmail,
+      },
+    });
+
+    if (!post) return next(new Error("Post not found."));
+
+    //// Delete img and remove folder
+    const folder = `${process.env.CLOUDINARY_FOLDER}/Posts/${post.draftId}`;
+    await cloudinary.uploader.destroy(post.img_publicId as string);
+    await cloudinary.api.delete_folder(folder);
 
     const deletedPost = await prisma.post.delete({
       where: {
@@ -61,6 +78,7 @@ export async function deletePost(
       })
     );
   } catch (error) {
+    console.log("Error in delete post : ", error);
     return next(new Error(handlePrismaError(error).message));
   }
 }
